+++ bgstally/constants.py
@@ -33,9 +33,10 @@ class UpdateUIPolicy(Enum):
 
 class DiscordChannel(Enum):
     BGS = 0
-    FLEETCARRIER_MATERIALS = 1
-    FLEETCARRIER_OPERATIONS = 2
-    THARGOIDWAR = 3
+    CMDR_INFORMATION = 1
+    FLEETCARRIER_MATERIALS = 2
+    FLEETCARRIER_OPERATIONS = 3
+    THARGOIDWAR = 4
 
 
 class MaterialsCategory(Enum):